Commands should work, missing generate
